Move `<top-bar>` component to src/sidebar/components
