+++ src/sidebar/components/top-bar.js
